summarize_package_results: avoid dplyr::join_by
join_by was introduced in dplyr 1.1.0, so at the very least
mpn.scorecard should declare a version constraint.  However, dplyr
1.1.0 (or actually vctrs underneath [*]) has a numeric_version sorting
bug, so the MPN toolchain won't upgrade to dplyr until a vctrs version
with the fix makes it to MPN.

Use the older, simpler syntax for compatibility with dplyr versions
before 1.1.0.

[*] fixed by 793d0282 (Implement a `numeric_version` equality proxy,
    2023-02-08)