coverage: do not catch warnings when running covr
add_coverage() specifies a warning handler for the tryCatch around its
covr run.  As a result, a warning halts execution.  In the best case
scenario, the warning is related to a downstream error that would have
been signaled, and that warning is captured in place of the error [*].
In the worst case, a warning can stop the execution of a run where an
error would not have been signaled.

Drop the warning handler to avoid this.  Note that this leads to some
warnings leaking into the test output.  Don't bother suppressing those
because they'll go away when the next commit switches to calling covr
in a separate process.

[*] An example of this case is the test-create-extra-notes.R assertion
    that needs to be adjusted for this change.  The previously
    captured message was an upstream warning, not the actual error
    message.

Closes #11.