score_pkg: support specifying timeout for covr
For the MPN build, the covr run for a few packages is stalling.
Unlike rcmdcheck::rcmdcheck(), covr::package_coverage does not have an
option to kill its run after a certain period of time.

Support a timeout by running covr within a separate callr process.

This of course comes with some overhead, but, even in single-package
benchmarks, it looks like other factors dominate.  Here is the
microbenchmark output for three runs of score_pkg with nmrec 0.1.0 as
the input:

  # main (86b6d7141)
       min       lq     mean   median       uq      max neval
  52.36488 53.18738 54.03379 54.00988 54.86825 55.72661     3

  # this commit
       min      lq     mean   median      uq      max neval
  52.90185 52.9272 54.62041 52.95255 55.4797 58.00684     3

Closes #5.