test-summarize-package-results.R: drop a magic number
After calling summarize_package_results() with result_dirs_select, a
test-summarize-package-results.R test checks the dimensions of the
result.

Assert that the number of rows matches the length of
result_dirs_select rather than a hard-coded number so that this test
doesn't need to be updated each time result_dirs_select (defined in
tests/testthat/setup.R) is updated.