create_extra_notes: handle 'no testable functions' case
create_extra_notes() detects that covr had an error by checking if
test_coverage is NA.  Doing so is incorrect because add_coverage()
also sets test_coverage to NA when there are no testable functions, in
which case covr does not error.  As a result, create_extra_notes()
will fail for these cases when trying to extract the error message.

Update create_extra_notes() to distinguish between the "covr error"
and "no testable function" edge cases.

Note that adding pass_no_functions case to the scored packages in
setup.R means that, in addition to the new test, an existing test in
test-render-scorecard.R serves as a regression test for these changes.

Closes #17.