tests: reduce assumptions about test case positions
tests/testthat/setup.R scores a set of packages that is used in many
of the test files.  Tests often index the list of results by position,
which 1) makes it hard for someone reading a test to know which case
is being tested and 2) means someone updating the package set needs to
adjust all the downstream tests if the positions change.

To make this setup more readable and less brittle, add names to
result_dirs_select and update the tests to index by name.