format_appendix: render failed coverage verbatim
For covr failures, format_appendix() includes the error message in the
test_coverage column of the flextable.  This contains arbitrary text
so, as with the 'R CMD check' cases described in the previous commit,
this may trigger a LaTeX compilation failure, with the added
complication that flextable also has logic to escape LaTeX code.

Even ignoring the compilation failures, rendering the failure within
table leads to a confusing result given that the error is not
"test_coverage" and no other results are in the table.

Only use a flextable if there are actually results to put in the
table.  For the "no testable function" case, just display the note as
regular text.  For the covr failure case, which may have arbitrary
text, display the error as verbatim text, following what was done for
the check output in the previous commit.

Re: #18