format_appendix: render check output verbatim
To render 'R CMD check' output, format_appendix() replaces each
backslash with two and then includes the output directly in the Rmd
document.

For most cases, the output is markdown that can be successfully
rendered (Rmd -> md -> tex -> pdf).  A handful of characters need to
be escaped at the LaTeX level ("&%$#_{}~^\"), but pandoc takes care of
all but one.  The exception is a single blackslash, which pandoc will
treat as a line break.  By doubling it, format_appendix() gets pandoc
to produce the desired "\textbackslash{}".

Even with the backslash escape, we're hitting into two cases (for all
the packages on MPN) where the check output makes the LaTeX
compilation fail.  The core problem is that we can't expect pandoc to
escape arbitrary input into valid LaTeX code.  As a minimal example,
take the characters '\$}\$'.  format_appendix() would escape that to
'\\$}\\$'.  pandoc would convert LaTeX code that has a backslash and
then inline math:

  \textbackslash{}\(}\\\)

That'd fail to compile with "Extra }, or forgotten $".

In addition to this, there's a more cosmetic issue: format_appendix()
is dumping the 'R CMD check' output to be rendered as markdown.  The
leading "* " in the sections are treated as bullets.  For sections
with content that span multiple lines, that renders poorly because
everything is collapsed to a single line.

Fix both the above issues by rendering this text as verbatim.  Note
that spverbatim is used because, unlike verbatim, it will wrap long
lines rather than cutting them off.

There are likely still other cases that can trip of the LaTeX
compilation.  For example, at least testing directly from an Rmd, a
vertical tab in the Rmd source will make xelatex failed with "text
line contains an invalid character", but that was a problem for the
old approach too.  Let's hold off on dealing with any remaining cases
until we hit into them.

Re: #18