DESCRIPTION: drop packages from Imports that aren't used in R/
The 'R CMD check' for mpn.scorecard has the following note [*]:

    Namespaces in Imports field not imported from:
       ‘here’ ‘markdown’ ‘tidyr’
      All declared Imports should be used.

here is only used in the top-level update-readme.R, not in any R/
code, so move it to pkgr.yml.

tidyr is only used in test code, so move it to Suggests.

markdown isn't used at all, so just remove it.

[*] This isn't flagged by the default `devtools::check()` (which
    passes --as-cran) until R 4.2 (see revision 81155).