render_scorecard: list dependencies and versions in appendix table
When building MPN or scoring a package with our internal helper
script, we can determine which dependencies were used because we know
the overall set of available packages.  That's not true in general, so
it'd be good to include information about the version of dependencies
that were used.

Calling sessionInfo() in the main session won't do what we want
because that will only capture loaded/attached packages for the
current session.  Both rcmdcheck and covr run in a subprocess.

rcmdcheck already captures the some session info via the
sessioninfo::session_info() package (using fairly complicated logic to
hook into the `R CMD check` subprocess).  That includes a data frame
of package names and versions for the package's dependencies [*].
Include that as a table in the appendix.

This new table falls under "System Info", so put it in that section.
The other table in that section is called "System Information", but
that table also contains info about the execution, so rename it to
"Execution and Machine Information".

[*] Unfortunately that call uses the default value for the
    `dependencies` argument, so suggested packages are not included.
    It's worth a feature request to rcmdcheck to support including
    suggested packages, but for now recording the hard dependency
    versions is better than nothing.

Closes #39.