test setup: always include dependency 'package with R code' cases
The pass_note test case adds an unused dplyr import to DESCRIPTION to
get a note.  Otherwise the test packages don't specify any packages
under 'Imports'.

For the current set of tests, that doesn't matter one way or the
other.  The next commit will add a dependency version table to the
appendix, so it's useful to have the tests cover both cases with and
without dependencies.

Add checkmate as a dependency for the cases that use the good R
snippet.