wrapi_text: don't process NA input
Calling format_traceability_matrix() for a package where some exports
in the namespace weren't mapped to an R/ file (as is the case with
mrgsolve) leads to an error following gh-57, which was included in the
mpn.scorecard 0.4.0 release.

The root of the error is wrapi_text() using an NA result in a
condition.  There's nothing for wrapi_text() to do with NA input, so
update it to return the NA input as is.

Closes #64.