ci: update R versions to cover latest Metworx
There are jobs for R 4.0.5 and 4.1.3 to match the two most recent R
versions on Metworx.  A new Metworx AMI (24.04) recently came out, and
its two latest R versions are 4.2.3 and 4.3.1.

Add jobs for those versions.

Retain the R 4.0.5 job to help flag incompatibles with the previous
AMI (22.09).  Drop the R 4.1.3 job because it's not likely to catch
something that wouldn't be caught by the other jobs.