wrapi_text: fix handling of successive parts that exceed width
wrapi_text() splits a string into pieces and then builds it back up,
inserting a newline before the next piece if it would lead to the
string exceeding the specified width.  The result is stored in the
`start` variable, which gets extended each time the width is exceeded.

To extend `start`, wrapi_text() appends the value of

  {pieces that did not exceed width}{sep}{newline}{piece that exceeded width}

wrapi_text() incorrectly assumes that there is at least one piece that
did _not_ exceed the width since the last time `start` was extended,
leading to a malformed result with the last piece repeated in an
incorrect position.

Update the logic to account for the fact that successive pieces can
exceed the specified width.