wrapi_text: fix indexing logic
wrapi_text() splits a string into pieces and then builds it back up,
inserting a newline before the next piece if it would lead to the
string exceeding the specified width.  The `shift` variable is set
whenever extending the string with the next piece would exceed the
maximum width.  It marks the last position that did _not_ exceed the
bound.

The logic for incrementing `shift`, however, is flawed.  It's
calculated by adding the current `shift` value to the current
iteration's index, resulting in skipped values and NAs (from
out-of-bounds indexing).

One way to correct the value would be to set `shift` to the index for
the current iteration.  Instead set it to the index for the _next_
iteration, redefining `shift` to be the first piece to consider when
extending the string.  Defining it this way lets us use `shift` rather
than `1 + shift` when indexing.