wrapi_text: fix spurious newlines
When handling a string (say "abcde") that exceeded the specified width
(say 2), wrapi_text() uses two values to put the final string
together:

 * start: a string that serves as frozen/good start.  It was
   calculated during the last iteration that exceeded the maximum
   width ("ab\nc", for the example above when splitting on each
   character and at the iteration for "e").

 * add: a string representing the characters after `start` through the
   character of the current iteration ("d\ne").

When wrapi_text() joins these two values, it includes a newline,
leading to a spurious newline between "c" and "d".

Drop the newline, leaving just the original character that was used to
split the string ("" in the example above).