scorecard-template.Rmd: fix default parameter values
The scorecard template uses values like "r NULL" and "`r NA`" for the
parameter defaults.  If these values aren't specified during the
rmarkdown::render(), they come through as strings (e.g., "r NULL" and
"`r NA`"), not the intended R objects.

At the moment, this doesn't matter because all the values are supplied
to the rmarkdown::render() call.  But that will change when support
for externally scored packages is added and dep_versions_df is left
unspecified.

Set the defaults to all these values to null, allowing downstream R
code to check whether the parameter is specified with is.null().