test-format-report.R: simplify an assertion
The test of format_scores_for_render() constructs the set of expected
criteria from the object read from *.scorecard.json, tweaking for the
renames of the testing criteria.  This makes it difficult for a reader
to know what's expected.

Reduce the amount of code logic in the test by replacing this logic
with a hard-coded list of expected criteria.