render_scorecard: support externally scored packages
With recent commits, all of the formatting functions have been
decoupled, at least conditionally, from the assumption that it is an R
package that's being scored.  Define a set of files that any type of
package can use to render a scorecard.

The structure of the results directory for external scores is similar
to that of the directory for "internal" scores.  check.txt replaces
check.rds, coverage.json replaces covr.rds, and matrix.yaml replaces
export_doc.rds.

For scorecard.json, an external _could_ write that file.  But doing so
is awkward because that file includes overall scores derived from
other scores in the file.  To compute those, the external scorer would
need to either duplicate the scoring logic or depend on mpn.scorecard.

So instead have the external scorer produce parts of scorecard.json,
and make mpn.scorecard responsible for stitching those parts together
at render time.

A notable limitation of the external input is that the check values
must be 0 or 1; 0 through 1, like is done for R packages, is not
supported.  There's a number of ways we could add support for this,
but punt on that given that all our initial non-R packages of interest
need only a pass/fail result.