format_appendix: rename 'R Script' column to 'Code File'
An upcoming commit will add support for externally scored non-R
packages, in which case 'R Script' doesn't work for the column name.

We could switch from 'R script' to something else at runtime base on
the type of package being scored.  However, 'Code File' seems
appropriate enough for the coverage result of any type of package, so
go with that to avoid unnecessary branching in the code and divergence
in the report.