map_answer: make including check score optional
The 'R CMD check' score factors in the numbers of
notes/warnings/errors in addition to whether the tests passed or
failed, so it's a score in the 0 to 1 range.  However, package types
other than R will likely only have a pass/fail, in which case it
doesn't make sense to display the specific score.