render: don't assume coverage values are already rounded
format_appendix() and format_scores_for_render() rely on the coverage
values coming in rounded.  An upcoming commit will allow rendering a
scorecard for non-R packages, with coverage values coming from an
external source.  In that case, we can't expect the values to come in
rounded, and, in any case, it makes sense for rendering code to be
responsible for the rendered format.

With this change, the values are always rendered to two decimal places
(e.g., "100.00%" instead of "100%").  That has the advantage of nicer
alignment in the table.