render_scorecard: drop note about traceability matrix file
This note isn't relevant for the user.  The *.export_doc.rds file is
not something that they should ever write without the help of
make_traceability_matrix(), and we do not advertise the underlying
structure.

And with upcoming support for externally scored package, a
user-defined matrix can come in through a YAML file.