summary: guard against external scores
render_scorecard_summary() and summarize_package_results() have not
been reworked to support external scores.  We use these functions for
MPN.  For non-R packages, we don't yet have a need for these
functions.

Document that these functions don't accept external scores, and add a
custom guard so that a more informative error is given if any results
directory with external scores is passed.