format_appendix: replace empty coverage table with note
External scorers are responsible for providing the *.coverage.json,
and that file can have an overall coverage score but not per file
scores.  If there are no rows in the coverage table, replace the table
with a note to make it clear that this is expected rather than a
rendering error.

Thanks to @seth127 for the suggestion.

Re: https://github.com/metrumresearchgroup/mpn.scorecard/pull/70#discussion_r1712232315