render_scorecard: fix recent regression in relative path handling
As aa2cb8f (render_scorecard: move preparation of params to dedicated
function, 2024-08-06), a function call instead of an inline list(...)
is passed to rmarkdown::render() as the params argument.  That change
breaks the handling of relative paths for results_dir because render()
evaluates the params argument after changing the working directory.

Move the function call outside of the rmarkdown::render() call to fix
the handling of relative paths.