templates: maintain aspect ratio when inserting title logo
Both a width and height is specified when inserting the logo above the
title, changing the aspect ratio and leading to noticeable distortion.
Avoid this by specifying only the width.

Increase the width a bit to push the height a little closer to what it
was before this change.  (A width of 3.6 cm would be needed to
maintain a height of 2 cm.)