templates: speed up rendering by using smaller png for header logo
Rendering scorecards on Metworx 25-02.00.00 has become slower compared
to rendering on Metworx 24-04.03.01 (e.g., rendering nmrec's scorecard
has gone from around 6 seconds to around 11 seconds).

The primary slowdown comes from the xelatex call.  In particular,
inserting the page header logo appears slower.  Perhaps this is due to
a change in XeLaTex (TeX Live 2023 versus Tex Live 2024) or an
underlying library (e.g., libpng).

The slowdown can be triggered by feeding xelatex a minimal tex file
(with only the fancyhdr setup and blank sections, for a total of 8
headers/pages) and can be decreased by using a smaller image (100x55
versus 2541x1401).  Here are timings (via hyperfine):

  25-02.00.00

    Benchmark 1: xelatex logobig
      Time (mean ± σ):      3.625 s ±  0.008 s    [User: 3.409 s, System: 0.233 s]
      Range (min … max):    3.612 s …  3.644 s    10 runs

    Benchmark 2: xelatex logosmall
      Time (mean ± σ):     662.7 ms ±   7.2 ms    [User: 575.2 ms, System: 105.6 ms]
      Range (min … max):   656.3 ms … 681.5 ms    10 runs

    Summary
      'xelatex logosmall' ran
        5.47 ± 0.06 times faster than 'xelatex logobig'

  24-04.03.01

    Benchmark 1: xelatex logobig
      Time (mean ± σ):     952.2 ms ±   8.5 ms    [User: 835.9 ms, System: 133.4 ms]
      Range (min … max):   944.6 ms … 974.9 ms    10 runs

    Benchmark 2: xelatex logosmall
      Time (mean ± σ):     566.4 ms ±   5.5 ms    [User: 471.8 ms, System: 111.8 ms]
      Range (min … max):   560.7 ms … 574.6 ms    10 runs

    Summary
      'xelatex logosmall' ran
        1.68 ± 0.02 times faster than 'xelatex logobig'

Notably the ~5x slowdown on 25-02 increases with the number of pages
and doesn't happen under pdflatex.

Switch to the smaller image to decrease the 25-02 render times.  There
should be no downside to doing so because 100x55 is sufficient
resolution for the small logo inserted into the header.